FPU fixes
Normalisation, always normalise as a double precision ( 53 bits ) .

Rounding and combining. Correct bits being used for single precision rounding.single precision.

Square root, use normalisation, instead of fastnormal.

Majority of glitches in OUTRUN demo now fixed.

Continuation of porting of QUAKE. Now loads the initial PAK0.PAK file and starts parsing.

+++ RV64/SOFTWARE/c/mc1-quake/WinQuake/common.h
@@ -8,7 +8,7 @@ of the License, or (at your option) any later version.
 
 This program is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
 
 See the GNU General Public License for more details.
 
@@ -63,7 +63,7 @@ void InsertLinkAfter (link_t *l, link_t *after);
 // (type *)STRUCT_FROM_LINK(link_t *link, type, member)
 // ent = STRUCT_FROM_LINK(link,entity_t,order)
 // FIXME: remove this mess!
-#define	STRUCT_FROM_LINK(l,t,m) ((t *)((byte *)l - (int)&(((t *)0)->m)))
+#define	STRUCT_FROM_LINK(l,t,m) ((t *)((byte *)l - (long)&(((t *)0)->m)))
 
 //============================================================================
